pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "python-backend"
version = "0.1.0"
description = "Python backend using Flask, pytest, Flasgger, and OpenAPI"
authors = [ { name="Your Name", email="you@example.com" } ]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask==3.0.3",
    "pytest==8.3.3",
    "python-dotenv==1.0.1",
    "flasgger==0.9.7.1",
    "pyyaml==6.0.2",
    "openapi-spec-validator==0.7.1"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]
